fetch ImportSettingsData
